+++ PapaCheck.Web/js/big-screen.js
@@ -1182,7 +1182,6 @@ async function redeemItem(itemId) {
     updateShopPage();
     showToast('兑换成功！');
     Voice.speak('兑换成功！');
-    Voice.speak('奖励箱有新奖励，快去看看吧');
   } finally {
     _redeemingItem = false;
   }